# tests/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mik API",
        "description": "\u7c73\u5ba2\u4e4b\u5bb6API",
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/shops": {
            "get": {
                "tags": [
                    "Shop"
                ],
                "summary": "\u83b7\u53d6\u5341\u4e2a\u7ec8\u7aef,\u6216\u8005\u641c\u7d22\u7ec8\u7aef",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u641c\u7d22\u9910\u5385"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shop"
                ],
                "summary": "\u6dfb\u52a0\u65b0\u7ec8\u7aef",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "distributor_id",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "area_id",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6dfb\u52a0\u65b0\u9910\u5385"
                    }
                }
            }
        },
...
        "/test/withdraw": {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "\u6d4b\u8bd5\u63d0\u73b0",
                "parameters": [
                    {
                        "name": "openid",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "amount",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u670d\u52a1\u5458\u6570\u636e"
                    }
                }
            }
        },
...
    },
...
}